+++ src/pages/HomePage/HomePage.styled.tsx
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const HomeWrapper = styled.div`
+  background-color: #7e7dfd;
+  @media screen and (min-width: 992px) {
+    div {
+      display: flex;
+      height: 100vh;
+    }
+  }
+`;
+
+// export const MediaHome = styled.div`
+//   @media screen and (min-width: 992px) {
+//     div {
+//       display: flex;
+//       flex-direction: row;
+//     }
+//   }
+// `;
