+++ src/pages/SettingPage/SettingPage.styled.tsx
@@ -30,7 +30,7 @@ export const ModalBackground = styled.div`
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
+  background-color: var(--bg-blur);
   display: flex;
   justify-content: center;
   align-items: center;
